+++ src/Enemy.cc
@@ -1,6 +1,6 @@
 #include "Enemy.h"
 
-Enemy::Enemy(Texture* t, const SDL_Rect& rect, double health, int score)
+Enemy::Enemy(Texture* t, const SDL_Rect& rect, float health, int score)
     : GameObject(t, rect, health),
       score{score} {}
